Updated the layouts a lot to all have green backgrounds and white text including the spinners which required new layout files for them (I don't know if I needed two or not).
Also changed the text on a few strings to shorten them and upped the size of the text to 18sp on the settings TextViews

Fixed bug where the Play fragment would be created again and the song and buttons would reset to their defaults (pause button will say pause not resumed when recreated while paused and needs to be fix later).

Changed the song selection form shuffing the ArrayList and selecting the first song to selecting a random song form anywhere in the list. This fixed the bug where the Music fragment would be recreated and change order.

Have the song and artist data sent to the TextViews when the timer calls the UpdateUI method so it is recreated when the fragment is recreated (MOVE THIS TO THE FRAGMENT ALONG WITH ALL OTHER UI INFROMATION SO IT CAN BE CREATED IN THE STATE IT WILL BE SEEN NOT 'LAG' A HALF SEC OR SO).