Added log messages to most files to make debugging easier.
Fixed bug with the UI song artist and title not showing up on swiping the screen.

Hopefully fixed bug with the second timer not triggering the min change. Assuming since the min change if statement was an equals, there was a situation where it was not called at the min change and caused the equals to never be true. Changed the if statement to a > now.

Fixed bug where songs would crash the app if they were all played once already. Now resets that flag if its needed. Still an issue when none are selected though.

Will not allow the pause toggle now when the game is paused causing a weird button state mixup